refactor(core): extract ledger types to pure module (PRI-443 Phase 1)
